Refactor logging package: implement OTLP support and enhance Zap logger integration
